requirements.txt
fastapi
uvicorn[standard]
python-dotenv
motor
pydantic
pydantic-settings
langchain
langchain-core
langchain-community
langchain-groq
langchain-huggingface
sentence-transformers
faiss-cpu
pyjwt
bcrypt
pytest
httpx
python-multipart
